% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{add.asterisks}
\alias{add.asterisks}
\title{Add two asterisks to the data frame in the positions}
\usage{
add.asterisks(x, w)
}
\arguments{
\item{x}{a data frame}

\item{w}{matrix 'where' to add asterisks to}
}
\value{
a data frame
}
\description{
Add two asterisks to the data frame in the positions
}
