% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{commify}
\alias{commify}
\title{Take the input vector and create a comma-seperated string}
\usage{
commify(vec, use_and = TRUE, and_word = "and")
}
\arguments{
\item{vec}{Vector of character strings to use}

\item{use_and}{Add the word 'and' before the last string}

\item{and_word}{Word to use for "and" -- could be "et" for french}
}
\value{
A string
}
\description{
Take the input vector and create a comma-seperated string
}
\examples{
library(gfutilities)
commify(c("One", "two", "three", "four"))
}
