% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{file_addext}
\alias{file_addext}
\title{Make filenames with the supplied extensions. If the extensions are already in the filenames, return the filenames}
\usage{
file_addext(filename, ext)
}
\arguments{
\item{filename}{name for the files with or without the extension}

\item{ext}{extensions to use. May or may not include a dot (.)}
}
\value{
The filename(s) with the supplied extension(s)
}
\description{
Make filenames with the supplied extensions. If the extensions are already in the filenames, return the filenames
}
\examples{
library(gfutilities)
file_addext("hello_world_data", ".rds")
}
