% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{firstlower}
\alias{firstlower}
\title{Make the first letter in strings lowercase}
\usage{
firstlower(x)
}
\arguments{
\item{x}{A vector of strings}
}
\value{
the same vector back, with the first letter non-capitalized for each vector element
}
\description{
Make the first letter in strings lowercase
}
\examples{
library(gfutilities)
firstlower("Hello")
firstlower(c("Hello", "World"))
}
