% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{firstup}
\alias{firstup}
\title{Capitalize the first letter in strings}
\usage{
firstup(x)
}
\arguments{
\item{x}{A vector of strings}
}
\value{
the same vector back, with the first letter capitalized for each vector element
}
\description{
Capitalize the first letter in strings
}
\examples{
library(gfutilities)
firstup("hello")
firstup(c("hello", "world"))
}
