% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{get.align}
\alias{get.align}
\title{Create a character vector to use in an align argument for xtable()}
\usage{
get.align(num, first.left = TRUE, just = "r")
}
\arguments{
\item{num}{The number of columns}

\item{first.left}{If TRUE, set the first column to be left-justified}

\item{just}{The justification of all columns, unless first.left is TRUE.
"r", "l", or "c"}
}
\value{
A character vector to use in an align argument for xtable()
}
\description{
Create a character vector to use in an align argument for xtable()
}
\examples{
library(gfutilities)
library(xtable)
d <- data.frame(a = c(1,2,3), b = c(10,20,30))
print(xtable(d,
             caption = "The table caption",
             label = "tab:example",
             align = get.align(ncol(d), just = "c")),
      caption.placement = "top",
      include.rownames = FALSE,
      table.placement = "H",
      sanitize.text.function = function(x){x},
      size = latex.size.str(12, 11))
}
