% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_shade}
\alias{get_shade}
\title{Get an RGB string representing the color and opacity given}
\usage{
get_shade(color, opacity = 0)
}
\arguments{
\item{color}{A vector of R color strings or numbers}

\item{opacity}{An integer between 0 an 99 representing percentage opaque}
}
\value{
A vector of RGB strings of the specified colors and opacity
        Format of returned string is #RRGGBBAA
        where RR = red, a 2-hexadecimal-digit string
              GG = green, a 2-hexadecimal-digit string
              BB = blue, a 2-hexadecimal-digit string
              AA = alpha or opacity
}
\description{
Get an RGB string representing the color and opacity given
}
\examples{
library(gfutilities)
get_shade("blue", 1)
get_shade("salmon")
get_shade(c(1, 2, 3), 50)
get_shade(c("red", "blue", "yellow"), 20)
}
