% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{group_quantiles}
\alias{group_quantiles}
\title{Calculate arbitrary quantiles for a particular column by a grouping}
\usage{
group_quantiles(
  df = NULL,
  grp_cols = NULL,
  cols = NULL,
  probs = c(0.05, 0.25, 0.5, 0.75, 0.95),
  include_mean = TRUE
)
}
\arguments{
\item{df}{A [data.frame]}

\item{grp_cols}{A character vector of the column names to use for grouping the data}

\item{cols}{A character vector representing column names on which to perform the
calculations. They must all be in `df` or an error will be thrown}

\item{probs}{A vector of quantile probabilities to pass to [stats::quantile()]}

\item{include_mean}{If TRUE, include the mean in the output as the column `avg`}
}
\value{
A [data.frame] with a new column for each value in the `probs` vector
}
\description{
Calculate arbitrary quantiles for a particular column by a grouping
}
\examples{
library(rlang)
library(dplyr)
library(purrr)
df <- storms \%>\% filter(year \%in\% 2000:2005)
probs <- c(0.05, 0.25, 0.5, 0.75, 0.95)
q <- group_quantiles(df, grp_cols = c("year", "month", "hour"),
 cols = c("wind", "pressure"), probs = probs)
print(head(q, 5))
}
