% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{has_specials}
\alias{has_specials}
\title{Do strings contain special characters}
\usage{
has_specials(str, white = FALSE)
}
\arguments{
\item{str}{A vector of strings to check}

\item{white}{Include whitespace as special characters?}
}
\value{
A logical vector representing if the input strings contain special characters or not
}
\description{
Do strings contain special characters
}
\examples{
library(gfutilities)
has_specials("HelloWorld")
has_specials(c("hello!", "asd", "11`"))
}
