% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{hex_hash}
\alias{hex_hash}
\title{Create a HEX color based on the characters of the input string}
\usage{
hex_hash(str, first = TRUE)
}
\arguments{
\item{str}{Input string}

\item{first}{if TRUE, use the first three characters to generate the color. If FALSE,
use a random selection of three characters from the string to generate the color}
}
\value{
a string representing the color, e.g. "#FF00EE"
}
\description{
Create a HEX color based on the characters of the input string
}
\examples{
library(gfutilities)
hex_hash("Yelloweye Rockfish", FALSE)
}
