% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex.amp}
\alias{latex.amp}
\title{Create a string of ampersands seperated by spaces}
\usage{
latex.amp(n = 1)
}
\arguments{
\item{n}{The number of ampersands in the string}
}
\value{
A string of ampersands seperated by spaces. The string will have
one leading space and one trailing space
}
\description{
Create a string of ampersands seperated by spaces
}
\examples{
latex.amp(10)
}
