% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex.bold}
\alias{latex.bold}
\title{Place \\textbf macro around text for latex boldface}
\usage{
latex.bold(txt)
}
\arguments{
\item{txt}{The text}
}
\value{
The text surrounded by \\textbf{}
}
\description{
Place \\textbf macro around text for latex boldface
}
\examples{
latex.bold("Test")
}
