% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex.cline}
\alias{latex.cline}
\title{Create a latex string to draw a horizontal line across the columns specified}
\usage{
latex.cline(cols)
}
\arguments{
\item{cols}{A string representing the columns to span e.g. "1-5"}
}
\value{
A latex string to draw a horizontal line across the columns specified
}
\description{
Create a latex string to draw a horizontal line across the columns specified
}
\examples{
latex.cline("1-5")
}
