% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex.cmidr}
\alias{latex.cmidr}
\title{Create a latex string to draw a horizontal line across the columns specified}
\usage{
latex.cmidr(cols, trim = "r")
}
\arguments{
\item{cols}{A string representing the columns to span e.g. "1-5"}

\item{trim}{Tells latex to trim the line a bit so that if there are two lines
close together they don't touch in the middle (See booktabs package).
Can be "l" for left, "r" for right, or "lr" for both sides of line}
}
\value{
A latex string to draw a horizontal line across the columns specified
}
\description{
Create a latex string to draw a horizontal line across the columns specified
}
\examples{
latex.cmidr("1-5", "lr")
}
