% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.hline.R
\docType{data}
\name{latex.hline}
\alias{latex.hline}
\title{A horizontal line in latex properly escaped for R}
\format{
An object of class \code{character} of length 1.
}
\usage{
latex.hline
}
\description{
A horizontal line in latex properly escaped for R
}
\keyword{datasets}
