% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex.italics}
\alias{latex.italics}
\title{Place \\emph macro around text for latex italics}
\usage{
latex.italics(txt)
}
\arguments{
\item{txt}{The text}
}
\value{
The text surrounded by \\emph{}
}
\description{
Place \\emph macro around text for latex italics
}
\examples{
latex.italics("Test")
}
