% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex.math.bold}
\alias{latex.math.bold}
\title{Place \\mathbf macro and dollar signs around text for latex math boldface}
\usage{
latex.math.bold(txt)
}
\arguments{
\item{txt}{The text}
}
\value{
The text surrounded by \\mathbf{} and dollar signs
}
\description{
Place \\mathbf macro and dollar signs around text for latex math boldface
}
\examples{
latex.math.bold("Test")
}
