% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex.mcol}
\alias{latex.mcol}
\title{Place \\multicolumn macro around text}
\usage{
latex.mcol(ncol, just, txt)
}
\arguments{
\item{ncol}{The number of columns to span}

\item{just}{The justification e.g. "l", "c", or "r" for left, center, right}

\item{txt}{The text}
}
\value{
The given text with the latex \\multicolumn{} macro around it
}
\description{
Place \\multicolumn macro around text
}
\examples{
latex.mcol(5, "l", "This is a test column header")
}
