% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex.mlc}
\alias{latex.mlc}
\title{Create a string which has been glued together using multi-line-cell
macro for latex}
\usage{
latex.mlc(latex.vec, make.bold = TRUE, math.bold = FALSE)
}
\arguments{
\item{latex.vec}{The vector of strings to glue together}

\item{make.bold}{the \\textbf macro will be inserted unless math.bold is TRUE}

\item{math.bold}{if TRUE, the \\mathbf macro will be used}
}
\value{
A string which has been glued together using multi-line-cell
macro for latex
}
\description{
mlc must be defined as follows in the latex code:
\\newcommand{\\mlc}[2][c]{\\begin{tabular}[#1]{@{}c@{}}#2\\end{tabular}}
}
\examples{
latex.mlc(c("This", "is a", "test"), make.bold = TRUE)
}
