% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex.mrow}
\alias{latex.mrow}
\title{Place \\multirow macro around text}
\usage{
latex.mrow(nrow, just, txt)
}
\arguments{
\item{nrow}{The number of rows to span}

\item{just}{The justification e.g. "l", "c", or "r" for left, center, right}

\item{txt}{The text}
}
\value{
The given text with the latex \\multirow{} macro around it
}
\description{
Place \\multirow macro around text
}
\examples{
latex.mrow(5, "l", "This is a test row header")
}
