% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.nline.R
\docType{data}
\name{latex.nline}
\alias{latex.nline}
\title{A newline in latex properly escaped for R}
\format{
An object of class \code{character} of length 1.
}
\usage{
latex.nline
}
\description{
A newline in latex properly escaped for R
}
\keyword{datasets}
