% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex.paste}
\alias{latex.paste}
\title{Create a string comprised of each element in the vector with an
ampersand in between}
\usage{
latex.paste(vec)
}
\arguments{
\item{vec}{A vector of strings}
}
\value{
A string comprised of each element in the vector with an
ampersand in between. The string will have one leading and one
trailing space
}
\description{
Create a string comprised of each element in the vector with an
ampersand in between
}
\examples{
latex.paste(c("a", "b", "c"))
}
