% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex.size.str}
\alias{latex.size.str}
\title{Create a latex string for setting the given font size and space size}
\usage{
latex.size.str(fnt.size, spc.size)
}
\arguments{
\item{fnt.size}{Font size}

\item{spc.size}{Space between text size}
}
\value{
A latex string for setting the given font size and space size
}
\description{
Create a latex string for setting the given font size and space size
}
\examples{
latex.size.str(10, 11)
}
