% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex.subscr}
\alias{latex.subscr}
\title{Create a subscript in latex}
\usage{
latex.subscr(main.txt, subscr.txt)
}
\arguments{
\item{main.txt}{The main part of the text}

\item{subscr.txt}{The subscripted part of the text}
}
\value{
A latex string with subscripting
}
\description{
Create a subscript in latex
}
\examples{
latex.subscr("B", "0")
}
