% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex.supscr}
\alias{latex.supscr}
\title{Create a superscript in latex}
\usage{
latex.supscr(main.txt, supscr.txt)
}
\arguments{
\item{main.txt}{The main part of the text}

\item{supscr.txt}{The superscripted part of the text}
}
\value{
A latex string with superscripting
}
\description{
Create a superscript in latex
}
\examples{
latex.supscr("x", "2")
}
