% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex.under}
\alias{latex.under}
\title{Place \\underline macro around text for latex underlining}
\usage{
latex.under(txt)
}
\arguments{
\item{txt}{The text}
}
\value{
The text surrounded by \\underline{}
}
\description{
Place \\underline macro around text for latex underlining
}
\examples{
latex.under("Test")
}
