% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{list_string}
\alias{list_string}
\title{Create a string out of the strings in vec, by glueing together with commas
and placing 'and' before the last one}
\usage{
list_string(vec)
}
\arguments{
\item{vec}{Character vector}
}
\value{
A string with the items of vec glued together using commas, with
'and' place before the last one
}
\description{
Create a string out of the strings in vec, by glueing together with commas
and placing 'and' before the last one
}
\examples{
library(gfutilities)
list_string("One")
list_string(c("One", "two"))
list_string(c("One", "two", "three", "four"))
}
