% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{modpath}
\alias{modpath}
\title{Copy a Windows path from the clipboard, replace the backslashes
with forward slashes, and write the new path string back to the
clipboard}
\usage{
modpath()
}
\value{
Nothing
}
\description{
Copy a Windows path from the clipboard, replace the backslashes
with forward slashes, and write the new path string back to the
clipboard
}
