% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{number_to_word}
\alias{number_to_word}
\title{Convert a number into an English word}
\usage{
number_to_word(x, th = FALSE, cap_first = FALSE)
}
\arguments{
\item{x}{The number to change into English}

\item{th}{Make the word end in 'st', 'nd', 'rd', or 'th'}

\item{cap_first}{Capitalize the first letter of the resulting word}
}
\value{
English text representing the number given
}
\description{
Convert a number into an English word
}
\examples{
number_to_word(1)
number_to_word(999999, th = TRUE)
number_to_word(101, cap_first = TRUE)
}
