% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{rasterify_cols}
\alias{rasterify_cols}
\title{'Rasterify' column names. Change x and y column names into lower case or upper case}
\usage{
rasterify_cols(df, inv = FALSE)
}
\arguments{
\item{df}{data frame}

\item{inv}{If TRUE, change x and y column names to upper case}
}
\value{
modified data frame
}
\description{
'Rasterify' column names. Change x and y column names into lower case or upper case
}
