% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{remove.asterisks}
\alias{remove.asterisks}
\title{Remove any asterisks found in the given data frame x}
\usage{
remove.asterisks(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a list of 2, first is data frame without
asterisks, second is matrix of TRUE for for position
so that the asterisks can be reapplied after formatting.
}
\description{
Remove any asterisks found in the given data frame x
}
