% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{rich_colors}
\alias{rich_colors}
\title{Create a vector of n colors that are perceptually equidistant and in an
order that is easy to interpret}
\usage{
rich_colors(n, alpha = 1)
}
\arguments{
\item{n}{The number of colors to produce}

\item{alpha}{Transparency, from 0 (fully transparent) to 1 (opaque)}
}
\value{
A vector of n colors
}
\description{
Create a vector of n colors that are perceptually equidistant and in an
order that is easy to interpret
}
\examples{
library(gfutilities)
ncol <- 10
x <- rich_colors(ncol, 1)
plot(1:ncol, 1:ncol, col = x, pch = 20, cex = 5)
}
