% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{round_down_even}
\alias{round_down_even}
\title{Round down to the nearest even number}
\usage{
round_down_even(x, base = 2)
}
\arguments{
\item{x}{Number to round}

\item{base}{Base for rounding. If 2, even numbers.}
}
\value{
Number rounded down
}
\description{
Round down to the nearest even number
}
\examples{
round_down_even(13.1)
round_down_even(14.22)
}
