% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{round_list}
\alias{round_list}
\title{Round all values in an arbitrarily complex [list]. All character values
are ignored and returned as they were}
\usage{
round_list(lst = NULL, digits = 2, print_elem_names = FALSE)
}
\arguments{
\item{lst}{A [list] of arbitrary complexity}

\item{digits}{The number of decimal points to round all numeric values to}

\item{print_elem_names}{Logical. Print names of each list element to the screen}
}
\value{
A [list] in the same format as `lst` but with all values rounded to
`digits` decimal points
}
\description{
Round all values in an arbitrarily complex [list]. All character values
are ignored and returned as they were
}
\details{
This is a recursive function and therefore can have an arbitrary
nesting of lists.
}
