% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{summarize_quants}
\alias{summarize_quants}
\title{Calculate arbitrary quantiles for particular column(s) by group}
\usage{
summarize_quants(
  df = NULL,
  grp_cols = NULL,
  cols = NULL,
  probs = c(0.05, 0.25, 0.5, 0.75, 0.95),
  include_mean = TRUE,
  check_args = TRUE
)
}
\arguments{
\item{df}{A \link{data.frame}}

\item{grp_cols}{A character vector of the column names to use for grouping
the data}

\item{cols}{A character vector representing column names on which to perform
the calculations}

\item{probs}{A vector of quantile probabilities to pass to \link[stats:quantile]{stats::quantile}}

\item{include_mean}{If \link{TRUE}, include the mean in the output as the column
ending in \verb{*_avg}}

\item{check_args}{If \code{TRUE}, run the code that verifies the argument
constraints. If \code{FALSE} skip the tests. You may want to choose \code{FALSE} for
situations when this function is placed in a loop to be run many times
for speed savings reasons}
}
\value{
A \link{data.frame} with a new column for each value in the \code{probs}
vector and a column for the mean in \code{include_mean} is \code{TRUE}
}
\description{
Calculate arbitrary quantiles for particular column(s) by group
and create a summary \link{data.frame} which holds a column for each
unique combination of values of columns (\code{cols}) and quantile
values (\code{probs}).
}
\details{
Create a new \link{data.frame} which summarizes the one passed as \code{df}.
The summary \link{data.frame} will have a column for each grouping column
(\code{grp_cols}) and a set of columns for each column name (\code{cols}), where
the set of columns is one for each value of \code{probs}. New columns will
be compose of the name in \code{cols}, an underscore, and the value in \code{probs}.
}
