% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{f}
\alias{f}
\title{Format a value in a really nice way, including rounding and trailing zeroes}
\usage{
f(x, dec.points = 0)
}
\arguments{
\item{x}{The value to format}

\item{dec.points}{The number of decimal points}
}
\value{
A string representing the formatted value.
}
\description{
Format a value in a really nice way, including rounding and trailing zeroes
}
\details{
The value will be rounded to the number of decimal points specified,
and will contain trailing zeroes if necessary so that the string has exactly
the correct number of decimal points. Commas will be placed in between
thousands.
}
\examples{
f(10000)
f(999999.1, 3)
}
