% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex.perc}
\alias{latex.perc}
\title{Escape all instances of the percent symbol properly for latex}
\usage{
latex.perc(vec)
}
\arguments{
\item{vec}{Character vector}
}
\value{
The input vector with all instances of the percent symbol escaped
properly for latex
}
\description{
Escape all instances of the percent symbol properly for latex
}
\examples{
latex.perc(c("2.5\%", "50\%", "97.5\%"))
}
